requirements.txt
docopt
falcon>=1.0
gunicorn>=19.4.5
requests
pyyaml